Parking slot identification system.
This is a Parking slot identification system. Based on the street you are in, when you intend to park you
identify the free and available parking slots and book a slot of your choice ,then pay for the slot according to hours you intend to park  and a ticket is generated. This system is real time and it direct the driver
to the parking lot. It will be reserved also according to the date you enter as long as you've paid for the slot.